Refactor: move ensure_x to separate file credentials.py
